metadata/requirements.txt
geographiclib==2.1
geopy==2.4.1
pillow==12.0.0